Update H001071.yaml
After submission of ZIP codes, the second form does not render immediately. This was causing the fill_in action to fail.

I have implemented a find selector with a 1 second delay to allow the form to completely render before starting the fill. The find was placed after the Zip Submission (line 19).

After this change, the form submitted successfully on our end.